backend/submods/pyproject.toml
[project]
name = "gyza"
version = "0.2.0"
description = "Encrypt and embed secret messages into images"
authors = [{name = "jingfreecs"}]
dependencies = [
  "contourpy==1.3.0",
  "cycler==0.12.1",
  "fonttools==4.58.0",
  "importlib_resources==6.5.2",
  "kiwisolver==1.4.7",
  "matplotlib==3.9.4",
  "numpy==2.0.2",
  "packaging==25.0",
  "pillow==11.2.1",
  "pycryptodome==3.23.0",
  "pyparsing==3.2.3",
  "python-dateutil==2.9.0.post0",
  "six==1.17.0",
  "zipp==3.21.0"
]

requires-python = ">=3.7"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
